Handle quit cleanup and cancel overlapping captures (#56)

+++ ScreenCapture/App/MenuBarController.swift
@@ -220,7 +220,11 @@ class MenuBarController: NSObject {
     }
 
     @objc private func quitApp() {
-        NSApp.terminate(nil)
+        if let appDelegate = NSApp.delegate as? AppDelegate {
+            appDelegate.requestQuit()
+        } else {
+            NSApp.terminate(nil)
+        }
     }
 
     @objc private func recordingDidStart() {
